Completed Sherlock and the Angry Beast.
Moved files around so that this repository is about programming
challenges rather than specifically HackerRank.

The solution to Sherlock and the Beast is not ideal; there must be a
mathematical solution to this which will solve it in constant time, but
I wanted to try out Python methods so I did it this way.